This commit introduces a new environment for Csharp/.NET programming language.
This environment holds in general the concepts and the naming conventions of the Java environment

So far the YetiCsharpInitializer class gets via sockets the types , all the constructors and the methods of the C# assemblies under test with their parameters. 

The aim of the current implementation is to store them at the same place where the Java environment stores them in order to have as less changes to the rest of the code.

It works only with constructors and methods that have as parameters primitives and objects of types(i.e. classes) that exist in the assemblies under test. It does not work for generic arguments arguments or arrays.

Probably as it is implemented now the generation of test cases is possible but this is not tested and further consideration is needed for that part.

git-svn-id: https://svn.origo.ethz.ch/yeti/trunk@31 7f4fa506-1231-44ca-96bd-ce53b71a5133