One check is removed and one check is added.
Check is removed from the addition of duplicate value to the vector because if a fault value is coming again and again and therefore it should have more chances to be picked so if we have more occurances of it in the vector than it means more chances of being picked up.

Check is added on yeti strategy so that only those values are added which are in that type e.g. for char datatype the value range is 0 to 65535. Only values between this range will be added to the vector off-course after conversion to first char and then Character object.

git-svn-id: https://svn.origo.ethz.ch/yeti/trunk@231 7f4fa506-1231-44ca-96bd-ce53b71a5133