DSS option is changed to DSSR where R stands for Random.
Since at the moment we are not interested in the nature of interesting values for a program therefore the code associated with DSSR to log the interesting values to the screen and text file is removed to increase performance.

DSSR is modified and now the list of interesting values is not emptied by the constructor.

git-svn-id: https://svn.origo.ethz.ch/yeti/trunk@251 7f4fa506-1231-44ca-96bd-ce53b71a5133